Changed balance from INT to BigDecimal

+++ src/main/java/models/Account.java
@@ -1,15 +1,17 @@
 package models;
 
+import java.math.BigDecimal;
+
 public class Account {
 
     // Attributes
     private int accountID;
     private int customerID;
     private int accountType;
-    private int balance;
+    private BigDecimal balance;
 
     // Constructor
-    public Account(int accountID, int customerID, int accountType, int balance) {
+    public Account(int accountID, int customerID, int accountType, BigDecimal balance) {
         this.accountID = accountID;
         this.customerID = customerID;
         this.accountType = accountType;
@@ -26,7 +28,7 @@ public class Account {
     public int getAccountType(){return accountType;}
     public void setAccountType(int accountType){this.accountType = accountType;}
 
-    public int getBalance(){return balance;}
-    public void setBalance(int balance){this.balance = balance;}
+    public BigDecimal getBalance(){return balance;}
+    public void setBalance(BigDecimal balance){this.balance = balance;}
 
 }
